@@ -2,6 +2,7 @@
   "name": "lnreader",
   "version": "2.0.7",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev:android": "pnpm run generate:env:debug && react-native run-android --appIdSuffix \"debug\" --active-arch-only",
     "dev:ios": "react-native run-ios",
@@ -51,8 +52,8 @@
     "@react-native-google-signin/google-signin": "^16.0.0",
     "@react-native-vector-icons/common": "^12.4.0",
     "@react-native-vector-icons/material-design-icons": "^12.4.0",
-    "@react-native/codegen": "^0.81.5",
-    "@react-native/gradle-plugin": "^0.81.5",
+    "@react-native/codegen": "^0.82.1",
+    "@react-native/gradle-plugin": "^0.82.1",
     "@react-navigation/bottom-tabs": "^7.8.6",
     "@react-navigation/native": "^7.1.21",
     "@react-navigation/native-stack": "^7.8.0",
@@ -81,8 +82,8 @@
     "lodash-es": "^4.17.21",
     "lottie-ios": "^3.5.0",
     "protobufjs": "^7.5.4",
-    "react": "19.1.0",
-    "react-native": "^0.81.5",
+    "react": "19.1.1",
+    "react-native": "^0.82.1",
     "react-native-background-actions": "^4.0.1",
     "react-native-device-info": "^14.1.1",
     "react-native-draggable-flatlist": "^4.0.3",
@@ -99,6 +100,7 @@
     "react-native-saf-x": "^2.2.3",
     "react-native-safe-area-context": "^5.6.2",
     "react-native-screens": "^4.18.0",
+    "react-native-sha256": "^1.4.10",
     "react-native-shimmer-placeholder": "^2.0.9",
     "react-native-tab-view": "^4.2.0",
     "react-native-url-polyfill": "^2.0.0",
@@ -116,24 +118,31 @@
     "@react-native-community/cli": "^20.0.2",
     "@react-native-community/cli-platform-android": "^20.0.2",
     "@react-native-community/cli-platform-ios": "^20.0.2",
-    "@react-native/babel-preset": "^0.81.5",
-    "@react-native/eslint-config": "^0.81.5",
-    "@react-native/metro-config": "^0.81.5",
-    "@react-native/typescript-config": "^0.81.5",
+    "@react-native/babel-preset": "^0.82.1",
+    "@react-native/eslint-config": "^0.82.1",
+    "@react-native/metro-config": "^0.82.1",
+    "@react-native/typescript-config": "^0.82.1",
+    "@testing-library/react-native": "^13.3.3",
     "@types/color": "^4.2.0",
     "@types/jest": "^30.0.0",
     "@types/lodash-es": "^4.17.12",
     "@types/node": "^24.10.1",
     "@types/react": "~19.1.17",
     "@types/sanitize-html": "^2.16.0",
+    "@typescript-eslint/eslint-plugin": "^8.48.1",
+    "@typescript-eslint/parser": "^8.48.1",
     "babel-plugin-module-resolver": "^5.0.2",
     "babel-plugin-react-compiler": "19.1.0-rc.3",
-    "eslint": "^8.57.1",
-    "husky": "^7.0.4",
-    "jest": "^29.7.0",
+    "eslint": "^9.39.1",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^7.0.1",
+    "eslint-plugin-react-native": "^5.0.0",
+    "husky": "^9.1.7",
+    "jest": "^30.2.0",
     "lint-staged": "^12.5.0",
-    "prettier": "2.8.8",
+    "prettier": "3.7.4",
     "react-native-dotenv": "^3.4.11",
+    "react-test-renderer": "19.1.1",
     "typescript": "~5.9.3"
   },
   "lint-staged": {
