@@ -2,7 +2,7 @@ const ReactCompilerConfig = {
   target: '19',
 };
 
-module.exports = function (api) {
+export default function (api) {
   api.cache(true);
   return {
     presets: ['module:@react-native/babel-preset'],
