@@ -1,7 +1,10 @@
 const fs = require('fs');
 const path = require('path');
 
-const srcPath = path.resolve(__dirname, '../src/services/authoritative-voice-map.ts');
+const srcPath = path.resolve(
+  __dirname,
+  '../src/services/authoritative-voice-map.ts',
+);
 const outMd = path.resolve(__dirname, '../docs/english-high-voices.md');
 const outCsv = path.resolve(__dirname, '../docs/english-high-voices.csv');
 
@@ -11,7 +14,9 @@ const content = fs.readFileSync(srcPath, 'utf8');
 const startToken = 'export const AUTHORITATIVE_VOICE_MAP';
 const startIndex = content.indexOf(startToken);
 if (startIndex === -1) {
-  console.error('Could not locate AUTHORITATIVE_VOICE_MAP in authoritative-voice-map.ts');
+  console.error(
+    'Could not locate AUTHORITATIVE_VOICE_MAP in authoritative-voice-map.ts',
+  );
   process.exit(1);
 }
 
@@ -35,14 +40,19 @@ for (; i < content.length; i++) {
 }
 
 if (depth !== 0) {
-  console.error('Failed to locate matching closing brace for AUTHORITATIVE_VOICE_MAP');
+  console.error(
+    'Failed to locate matching closing brace for AUTHORITATIVE_VOICE_MAP',
+  );
   process.exit(1);
 }
 
 let rawObj = content.slice(braceStart, i + 1);
 
 // Remove any getDisplayDisplayName functions to avoid eval issues
-rawObj = rawObj.replace(/getDisplayDisplayName\s*:\s*function\([^\)]*\)\s*\{[\s\S]*?\},?/g, '');
+rawObj = rawObj.replace(
+  /getDisplayDisplayName\s*:\s*function\([^\)]*\)\s*\{[\s\S]*?\},?/g,
+  '',
+);
 
 let voiceMap;
 try {
@@ -52,28 +62,70 @@ try {
   process.exit(1);
 }
 
-const isEnglish = (v) => v.language && v.language.startsWith('en');
-const isHigh = (v) => v.quality === 'veryHigh' || v.quality === 'high';
+const isEnglish = v => v.language && v.language.startsWith('en');
+const isHigh = v => v.quality === 'veryHigh' || v.quality === 'high';
 
 const rows = [];
 for (const [key, val] of Object.entries(voiceMap)) {
   try {
     if (isEnglish(val) && isHigh(val)) {
-      rows.push({ key, id: val.id || '', displayName: val.displayName || '', language: val.language || '', region: val.region || '', name: val.name || '', quality: val.quality || '', platform: val.platform || '', altIds: (val.altIds || []).join('; '), nativeIds: (val.nativeIds || []).join('; ') });
+      rows.push({
+        key,
+        id: val.id || '',
+        displayName: val.displayName || '',
+        language: val.language || '',
+        region: val.region || '',
+        name: val.name || '',
+        quality: val.quality || '',
+        platform: val.platform || '',
+        altIds: (val.altIds || []).join('; '),
+        nativeIds: (val.nativeIds || []).join('; '),
+      });
     }
   } catch (e) {
     // ignore
   }
 }
 
-rows.sort((a,b)=> (a.language||'').localeCompare(b.language||'') || (a.region||'').localeCompare(b.region||'') || (a.name||'').localeCompare(b.name||''));
+rows.sort(
+  (a, b) =>
+    (a.language || '').localeCompare(b.language || '') ||
+    (a.region || '').localeCompare(b.region || '') ||
+    (a.name || '').localeCompare(b.name || ''),
+);
 
 // Write CSV
-const csvHeader = ['mapKey','id','displayName','language','region','name','quality','platform','altIds','nativeIds'];
+const csvHeader = [
+  'mapKey',
+  'id',
+  'displayName',
+  'language',
+  'region',
+  'name',
+  'quality',
+  'platform',
+  'altIds',
+  'nativeIds',
+];
 const csvLines = [csvHeader.join(',')];
 for (const r of rows) {
-  const escape = (s) => '"' + String(s).replace(/"/g,'""') + '"';
-  csvLines.push([r.key, r.id, r.displayName, r.language, r.region, r.name, r.quality, r.platform, r.altIds, r.nativeIds].map(escape).join(','));
+  const escape = s => '"' + String(s).replace(/"/g, '""') + '"';
+  csvLines.push(
+    [
+      r.key,
+      r.id,
+      r.displayName,
+      r.language,
+      r.region,
+      r.name,
+      r.quality,
+      r.platform,
+      r.altIds,
+      r.nativeIds,
+    ]
+      .map(escape)
+      .join(','),
+  );
 }
 fs.mkdirSync(path.dirname(outMd), { recursive: true });
 fs.writeFileSync(outCsv, csvLines.join('\n'));
@@ -82,11 +134,22 @@ fs.writeFileSync(outCsv, csvLines.join('\n'));
 const mdLines = [];
 mdLines.push('# English voices (quality = veryHigh / high)');
 mdLines.push('');
-mdLines.push('| Map Key | id | displayName | language | region | name | quality | platform | altIds | nativeIds |');
+mdLines.push(
+  '| Map Key | id | displayName | language | region | name | quality | platform | altIds | nativeIds |',
+);
 mdLines.push('|---|---|---|---|---|---|---|---|---|---|');
 for (const r of rows) {
-  mdLines.push(`| ${r.key} | ${r.id} | ${r.displayName} | ${r.language} | ${r.region} | ${r.name} | ${r.quality} | ${r.platform} | ${r.altIds} | ${r.nativeIds} |`);
+  mdLines.push(
+    `| ${r.key} | ${r.id} | ${r.displayName} | ${r.language} | ${r.region} | ${r.name} | ${r.quality} | ${r.platform} | ${r.altIds} | ${r.nativeIds} |`,
+  );
 }
 fs.writeFileSync(outMd, mdLines.join('\n'));
 
-console.log('Wrote', rows.length, 'English HQ/HIGH voices to', outMd, 'and', outCsv);
+console.log(
+  'Wrote',
+  rows.length,
+  'English HQ/HIGH voices to',
+  outMd,
+  'and',
+  outCsv,
+);
