@@ -49,7 +49,7 @@ const TrackSheet: React.FC<TrackSheetProps> = ({ bottomSheetRef, novel }) => {
     closeBottomSheet();
     setActiveTracker(tracker);
     setTrackSearchDialog(true);
-  }, []);
+  }, [closeBottomSheet]);
 
   const handleSetStatusDialog = useCallback(
     (tracker: TrackerMetadata) => {
